use contains intead of dupp + files renaming
